+++ src/test/java/com/zipcodewilmington/danny_do_better_exercises/mathutilities/TestMultiplication.java
@@ -42,7 +42,7 @@ public class TestMultiplication {
         short multiplier = 1;
         short expectedShort = 2;
         // : When
-        short actualShort = mathUtils.multiply(multiplicand, multiplier);
+        short actualShort = (short) mathUtils.multiply(multiplicand, multiplier);
         // : Then
         assertEquals(expectedShort, actualShort);
     }
@@ -53,7 +53,7 @@ public class TestMultiplication {
         byte multiplier = 14;
         byte expectedByte = -32;
         // : When
-        byte actualByte = mathUtils.multiply(multiplicand, multiplier);
+        byte actualByte = (byte) mathUtils.multiply(multiplicand, multiplier);
         // : Then
         assertEquals(expectedByte, actualByte);
     }
