+++ src/test/java/com/zipcodewilmington/danny_do_better_exercises/mathutilities/TestSubtraction.java
@@ -41,7 +41,7 @@ public class TestSubtraction {
         short difference = 16383;
         short expectedShort = 1;
         // : When
-        short actualShort = mathUtils.subtract(baseValue, difference);
+        short actualShort = (short) mathUtils.subtract(baseValue, difference);
         // : Then
         assertEquals(expectedShort,actualShort);
     }
@@ -52,7 +52,7 @@ public class TestSubtraction {
         byte difference = 64;
         byte expectedByte = -1;
         // : When
-        byte actualByte = mathUtils.subtract(baseValue, difference);
+        byte actualByte = (byte) mathUtils.subtract(baseValue, difference);
         // : Then
         assertEquals(expectedByte,actualByte);
     }
